# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DataExportList": {
        "properties": {
          "items": {
            "description": "The latest requests, newest first.",
            "items": {
              "$ref": "#/components/schemas/DataExportOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DataExportList",
        "type": "object"
      },
      "DataExportOut": {
        "properties": {
          "downloaded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downloaded At"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the emailed link stops working.",
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ready_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ready At"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "status": {
            "description": "`requested`: being built; `ready`: the link was emailed; `expired`: the link ran out; `failed`: it couldn't be built or emailed, ask again.",
            "enum": [
              "requested",
              "ready",
              "expired",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "requested_at",
          "ready_at",
          "expires_at",
          "downloaded_at"
        ],
        "title": "DataExportOut",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable, machine-readable error code, e.g. 'not_found'.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context, e.g. field validation errors.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable explanation; safe to show to users.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Correlates the response with server logs.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Verified At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_moderator": {
            "default": false,
            "description": "May use the moderation page and /api/v1/moderation.",
            "title": "Is Moderator",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "status": {
            "default": "active",
            "description": "`paused`: hidden from matching and new intros; chats and sign-in carry on.",
            "enum": [
              "active",
              "paused"
            ],
            "title": "Status",
            "type": "string"
          },
          "terms_accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Accepted At"
          },
          "terms_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Version"
          }
        },
        "required": [
          "id",
          "email",
          "created_at",
          "email_verified_at",
          "last_login_at",
          "terms_version",
          "terms_accepted_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Cynergi API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/data-exports": {
      "get": {
        "operationId": "list_data_exports_api_v1_me_data_exports_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExportList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          }
        },
        "summary": "My data download requests",
        "tags": [
          "account"
        ]
      },
      "post": {
        "description": "Builds the file in the background and emails a link that works for\n``DATA_EXPORT_LINK_HOURS``. The file has your account, profile, requests, intros,\nconnections and the messages you sent and received.",
        "operationId": "request_data_export_api_v1_me_data_exports_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "One is being built, or one was asked for recently (`data_export_recent`)."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Storage nearly full; try again later."
          }
        },
        "summary": "Ask for a copy of my data",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/me/data-exports/{export_id}/download": {
      "get": {
        "description": "Needs both your session and the token from the emailed link.",
        "operationId": "download_data_export_api_v1_me_data_exports__export_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Id",
              "type": "string"
            }
          },
          {
            "description": "From the email.",
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "description": "From the email.",
              "maxLength": 128,
              "minLength": 1,
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "The file, as a JSON attachment."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not one of your requests (`data_export_not_found`)."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Expired, not ready or a wrong token (`data_export_unavailable`)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download my data",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/me/pause": {
      "post": {
        "description": "Hidden from matching and can't receive new intros. Sign-in and existing chats carry\non; ``/me/resume`` undoes it.",
        "operationId": "pause_api_v1_me_pause_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Only an active account can be paused (`cannot_pause`)."
          }
        },
        "summary": "Pause my account",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/v1/me/resume": {
      "post": {
        "operationId": "resume_api_v1_me_resume_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The account isn't paused (`not_paused`)."
          }
        },
        "summary": "Resume my paused account",
        "tags": [
          "account"
        ]
      }
    },
...
  }
}